feat: #WB2-1057, transform editor in png to print mindmap + export function exporter
