+++ frontend/src/features/export-modal/useExportMindmap.ts
@@ -1,16 +1,6 @@
 import { useEffect, useState } from "react";
 
-import { Alert, useHotToast } from "@edifice-ui/react";
-import {
-  Designer,
-  ImageExporterFactory,
-  Exporter,
-  SizeType,
-  TextExporterFactory,
-  Mindmap,
-  // @ts-ignore
-} from "@edifice-wisemapping/editor";
-import { useTranslation } from "react-i18next";
+import { exporter } from "~/services/utils";
 
 type ExportFormat = "svg" | "jpg" | "png" | "mm" | "wxml";
 type ExportGroup = "image" | "mindmap-tool";
@@ -22,13 +12,10 @@ export const useExportMindmap = ({
   mapName: string;
   onSuccess: () => void;
 }) => {
-  const { t } = useTranslation();
-
   const [submit, setSubmit] = useState<boolean>(false);
   const [exportFormat, setExportFormat] = useState<ExportFormat>("svg");
   const [exportGroup, setExportGroup] = useState<ExportGroup>("image");
   const [zoomToFit, setZoomToFit] = useState<boolean>(true);
-  const { hotToast } = useHotToast(Alert);
 
   const handleOnSubmit = (): void => {
     setSubmit(true);
@@ -58,54 +45,9 @@ export const useExportMindmap = ({
     setExportFormat(defaultFormat);
   };
 
-  const exporter = async (formatType: ExportFormat): Promise<string> => {
-    let svgElement: Element | null = null;
-    let size: SizeType;
-    let mindmap: Mindmap;
-
-    const designer: Designer = globalThis.designer;
-    // exporting from editor toolbar action
-    if (designer) {
-      // Depending on the type of export. It will require differt POST.
-      const workspace = designer.getWorkSpace();
-      svgElement = workspace.getSVGElement();
-      size = { width: window.innerWidth, height: window.innerHeight };
-      mindmap = designer.getMindmap();
-    } else {
-      hotToast.error(t("mindmap.export.failed"));
-    }
-
-    let exporter: Exporter;
-    switch (formatType) {
-      case "png":
-      case "jpg":
-      case "svg": {
-        exporter = ImageExporterFactory.create(
-          formatType,
-          svgElement,
-          size.width,
-          size.height,
-          zoomToFit,
-        );
-        break;
-      }
-      case "wxml":
-      case "mm": {
-        exporter = TextExporterFactory.create(formatType, mindmap);
-        break;
-      }
-      default: {
-        const exhaustiveCheck: never = formatType as never;
-        throw new Error(`Unhandled color case: ${exhaustiveCheck}`);
-      }
-    }
-
-    return exporter.exportAndEncode();
-  };
-
   useEffect(() => {
     if (submit) {
-      exporter(exportFormat)
+      exporter(exportFormat, zoomToFit)
         .then((url: string) => {
           // Create hidden anchor to force download ...
           const anchor: HTMLAnchorElement = document.createElement("a");
